removed double long

+++ _printf.c
@@ -26,7 +26,7 @@ int no_struct(char c, int count, va_list argu)
 {
 	int j;
 	char *s;
-	long double binary;
+	int binary;
 
 	switch (c)/* A */
 	{
